Rewrite using jotai.
