+++ packages/commonwealth/client/scripts/views/menus/create_content_menu.tsx
@@ -13,7 +13,7 @@ import { CWMobileMenu } from '../components/component_kit/cw_mobile_menu';
 
 import { CWSidebarMenu } from '../components/component_kit/cw_sidebar_menu';
 import { useCommonNavigate } from 'navigation/helpers';
-import { useUser } from 'context/userContext';
+import { useUser } from 'atoms/userAtom';
 
 const getCreateContentMenuItems = (navigate): PopoverMenuItem[] => {
   const showSnapshotOptions =
