+++ packages/commonwealth/package.json
@@ -179,6 +179,7 @@
     "https-browserify": "^1.0.0",
     "ignore-loader": "^0.1.2",
     "jdenticon": "^2.1.1",
+    "jotai": "^2.0.3",
     "jquery": "^3.3.1",
     "jsdom": "^16.5.0",
     "jsdom-global": "^3.0.2",
